Move pieces into place, start converting to adapter API
