+++ .gitignore
@@ -1,0 +1,3 @@
+demo/main.js
+.DS_Store
+node_modules
