+++ demo/index.html
@@ -1,0 +1,8 @@
+<html>
+  <head>
+    <script language="javascript" src="main.js"></script>
+  </head>
+  <body>
+    <h1>Media Recorder Demo</h1>
+  </body>
+</html>
