+++ demo/src/main.js
@@ -1,0 +1,5 @@
+var Recorder = require('../../lib/media-recorder');
+
+console.log(Recorder);
+
+
