+++ lib/adapters.js
@@ -1,6 +1,16 @@
 'use strict';
 
+/*
+ * Adapters to mixin support for different recording backends.
+ *
+ * API:
+ * - #startRecording
+ * - #stopRecording
+ * - #startPlayback
+ * - #stopPlayback
+ */
+
 module.exports = {
-  'flash': require('adapters/flash'),
-  'recorderjs': require('adapters/recorderjs')
+  'swf': require('./adapters/swf'),
+  'recorderjs': require('./adapters/recorderjs')
 };
