+++ lib/adapters/swf/swfobject.js
