+++ package.json
@@ -1,9 +1,16 @@
 {
   "name": "media-recorder",
   "version": "0.0.1",
+  "scripts": {
+    "build:demo": "browserify demo/src/main.js -o demo/main.js"
+  },
   "main": "./lib/media-recorder.js",
   "dependencies": {
     "recorderjs": "git+https://github.com/chrisguttandin/Recorderjs.git",
     "underscore": "*"
+  },
+
+  "devDependencies": {
+    "browserify": "^11.1.0"
   }
 }
